fix(): 노선에 등록된 지하철역 제외시 예외처리 되지 않던 부분 수정

+++ src/main/java/wooteco/subway/line/application/LineService.java
@@ -11,6 +11,7 @@ import wooteco.subway.line.dto.SectionRequest;
 import wooteco.subway.line.exception.ExistLineColorException;
 import wooteco.subway.line.exception.ExistLineNameException;
 import wooteco.subway.line.exception.LineNotExistException;
+import wooteco.subway.line.exception.NotAbleToDeleteStationInLineException;
 import wooteco.subway.station.application.StationService;
 import wooteco.subway.station.domain.Station;
 
@@ -107,8 +108,10 @@ public class LineService {
     public void removeLineStation(Long lineId, Long stationId) {
         Line line = findLineById(lineId);
         Station station = stationService.findStationById(stationId);
+        if (sectionDao.findByStationId(stationId).isPresent()) {
+            throw new NotAbleToDeleteStationInLineException();
+        }
         line.removeSection(station);
-
         sectionDao.deleteByLineId(lineId);
         sectionDao.insertSections(line);
     }
